Crate decimal

Macros

d128
A macro to construct d128 literals.

Structs

Status
The status of the floating point context. Instead of faulting after every operation errors are added to this status. User code can check and/or clear the status fully or partially at specific points to check the validity of the computation. Each thread gets its own status. The status is initially empty.

d128
A 128-bit decimal floating point type.

Enums

Class

Rounding

Constants

CONVERSION_SYNTAX
Conversion syntax error.

DIVISION_BY_ZERO
Division by zero.

DIVISION_IMPOSSIBLE
Division impossible.

DIVISION_UNDEFINED
Division undefined.

INEXACT
The result is inexact.

INVALID_OPERATION
An invalid operation was requested.

OVERFLOW
Overflow.

UNDERFLOW
Underflow.