Struct decimal::d128

#[repr(C)]

pub struct d128 { /* fields omitted */ }

A 128-bit decimal floating point type.

Methods

impl d128

pub unsafe fn from_raw_bytes(bytes: [u8; 16]) -> d128

Creates a d128 from raw bytes. Endianess is host dependent.

pub fn to_raw_bytes(&self) -> [u8; 16]

Returns raw bytes for this d128. Endianess is host dependent.

pub fn get_status() -> Status

Returns the thread local status.

pub fn set_status(status: Status)

Sets the thread local status.

pub fn from_hex(s: &str) -> d128

Reads the hex binary representation from a string. This is the reverse of formatting with {:x}.

pub fn zero() -> d128

Returns the d128 representing +0.

pub fn infinity() -> d128

Returns the d128 representing +Infinity.

pub fn neg_infinity() -> d128

Returns the d128 representing -Infinity.

pub fn abs(self) -> d128

Returns the absolute value of self.

pub fn mul_add<O: AsRef<d128>>(self, a: O, b: O) -> d128

Calculates the fused multiply-add self × a + b and returns the result. The multiply is carried out first and is exact, so this operation has only the one, final, rounding.

pub fn logb(self) -> d128

Returns the adjusted exponent of self, according to IEEE 754 rules. That is, the exponent returned is calculated as if the decimal point followed the first significant digit (so, for example, if self were 123 then the result would be 2). If self is infinite, the result is +Infinity. If self is a zero, the result is –Infinity, and the DIVISION_BY_ZERO flag is set. If self is less than zero, the absolute value of self is used. If self is 1, the result is 0. NaNs are handled (propagated) as for arithmetic operations.

pub fn max<O: AsRef<d128>>(self, other: O) -> d128

If both self and other are numeric (not NaNs) this returns the larger of the two (compared using total ordering, to give a well-defined result). If either (but not both of) is a quiet NaN then the other argument is the result; otherwise NaNs are handled as for arithmetic operations.

pub fn min<O: AsRef<d128>>(self, other: O) -> d128

If both self and other are numeric (not NaNs) this returns the smaller of the two (compared using total ordering, to give a well-defined result). If either (but not both of) is a quiet NaN then the other argument is the result; otherwise NaNs are handled as for arithmetic operations.

pub fn next(self) -> d128

Returns the ‘next’ d128 to self in the direction of +Infinity according to IEEE 754 rules for nextUp. The only status possible is INVALID_OPERATION (from an sNaN).

pub fn previous(self) -> d128

Returns the ‘next’ d128 to self in the direction of –Infinity according to IEEE 754 rules for nextDown. The only status possible is INVALID_OPERATION (from an sNaN).

pub fn pow<O: AsRef<d128>>(self, exp: O) -> d128

The number is set to the result of raising self to the power of exp. Results will be exact when exp has an integral value and the result does not need to be rounded, and also will be exact in certain special cases, such as when self is a zero (see the arithmetic specification for details). Inexact results will always be full precision, and will almost always be correctly rounded, but may be up to 1 ulp (unit in last place) in error in rare cases. This is a mathematical function; the 106 restrictions on precision and range apply as described above, except that the normal range of values is allowed if exp has an integral value in the range –1999999997 through +999999999.

pub fn exp<O: AsRef<d128>>(self, exp: O) -> d128

The number is set to e raised to the power of exp. Finite results will always be full precision and inexact, except when exp is a zero or –Infinity (giving 1 or 0 respectively). Inexact results will almost always be correctly rounded, but may be up to 1 ulp (unit in last place) in error in rare cases. This is a mathematical function; the 106 restrictions on precision and range apply as described above.

pub fn ln(self) -> d128

The number is set to the natural logarithm (logarithm in base e) of self. self must be positive or a zero. Finite results will always be full precision and inexact, except when self is equal to 1, which gives an exact result of 0. Inexact results will almost always be correctly rounded, but may be up to 1 ulp (unit in last place) in error in rare cases. This is a mathematical function; the 106 restrictions on precision and range apply as described above.

pub fn log10(self) -> d128

The number is set to the logarithm in base ten of self. self must be positive or a zero. Finite results will always be full precision and inexact, except when self is equal to an integral power of ten, in which case the result is the exact integer. Inexact results will almost always be correctly rounded, but may be up to 1 ulp (unit in last place) in error in rare cases. This is a mathematical function; the 106 restrictions on precision and range apply as described above.

pub fn towards<O: AsRef<d128>>(self, other: O) -> d128

Returns the ‘next’ d128 to self in the direction of other according to proposed IEEE 754 rules for nextAfter. If self == other the result is self. If either operand is a NaN the result is as for arithmetic operations. Otherwise (the operands are numeric and different) the result of adding (or subtracting) an infinitesimal positive amount to self and rounding towards +Infinity (or –Infinity) is returned, depending on whether other is larger (or smaller) than self. The addition will set flags, except that if the result is normal (finite, non-zero, and not subnormal) no flags are set.

pub fn quantize<O: AsRef<d128>>(self, other: O) -> d128

Returns self set to have the same quantum as other, if possible (that is, numerically the same value but rounded or padded if necessary to have the same exponent as other, for example to round a monetary quantity to cents).

pub fn reduce(self) -> d128

Returns a copy of self with its coefficient reduced to its shortest possible form without changing the value of the result. This removes all possible trailing zeros from the coefficient (some may remain when the number is very close to the most positive or most negative number). Infinities and NaNs are unchanged and no status is set unless self is an sNaN. If self is a zero the result exponent is 0.

pub fn rotate<O: AsRef<d128>>(self, amount: O) -> d128

The result is a copy of self with the digits of the coefficient rotated to the left (if amount is positive) or to the right (if amount is negative) without adjusting the exponent or the sign of self. amount is the count of positions to rotate and must be a finite integer (with exponent=0) in the range -34 through +34. NaNs are propagated as usual. If self is infinite the result is Infinity of the same sign. No status is set unless amount is invalid or an operand is an sNaN.

pub fn scaleb<O: AsRef<d128>>(self, other: O) -> d128

This calculates self × 10other and returns the result. other must be an integer (finite with exponent=0) in the range ±2 × (34 + 6144), typically resulting from logb. Underflow and overflow might occur. NaNs propagate as usual.

pub fn compare<O: AsRef<d128>>(&self, other: O) -> d128

Compares self and other numerically and returns the result. The result may be –1, 0, 1, or NaN (unordered); –1 indicates that self is less than other, 0 indicates that they are numerically equal, and 1 indicates that self is greater than other. NaN is returned only if self or other is a NaN.

pub fn compare_total<O: AsRef<d128>>(&self, other: O) -> d128

Compares self and other using the IEEE 754 total ordering (which takes into account the exponent) and returns the result. No status is set (a signaling NaN is ordered between Infinity and NaN). The result will be –1, 0, or 1.

pub fn canonical(self) -> d128

Returns self ensuring that the encoding is canonical.

pub fn class(&self) -> Class

Returns the class of self.

pub fn classify(&self) -> FpCategory

Same as class() but returns std::num::FpCategory.

pub fn digits(&self) -> u32

Returns the number of significant digits in self. If self is a zero or is infinite, 1 is returned. If self is a NaN then the number of digits in the payload is returned.

pub fn is_canonical(&self) -> bool

Returns true if the encoding of self is canonical, or false otherwise.

pub fn is_finite(&self) -> bool

Returns true if self is neither infinite nor a NaN, or false otherwise.

pub fn is_integer(&self) -> bool

Returns true if self is finite and its exponent is zero, or false otherwise.

pub fn is_logical(&self) -> bool

Returns true if self is a valid argument for logical operations (that is, self is zero or positive, an integer (finite with a zero exponent) and comprises only zeros and/or ones), or false otherwise.

pub fn is_infinite(&self) -> bool

Returns true if the encoding of self is an infinity, or false otherwise.

pub fn is_nan(&self) -> bool

Returns true if self is a NaN (quiet or signaling), or false otherwise.

pub fn is_negative(&self) -> bool

Returns true if self is less than zero and not a NaN, or false otherwise.

pub fn is_normal(&self) -> bool

Returns true if self is a normal number (that is, is finite, non-zero, and not subnormal), or false otherwise.

pub fn is_positive(&self) -> bool

Returns true if self is greater than zero and not a NaN, or false otherwise.

pub fn is_signaling(&self) -> bool

Returns true if self is a signaling NaN, or false otherwise.

pub fn is_signed(&self) -> bool

Returns true if self has a minus sign, or false otherwise. Note that zeros and NaNs may have a minus sign.

pub fn is_subnormal(&self) -> bool

Returns true if self is subnormal (that is, finite, non-zero, and with magnitude less than 10-6143), or false otherwise.

pub fn is_zero(&self) -> bool

Returns true if self is zero, or false otherwise.